# Speedup `_setattr` usage and fix slight performance regressions