--- a/changelog.d/991.change.rst
+++ b/changelog.d/991.change.rst
@@ -1,0 +1,1 @@
+Fix slight performance regression in classes with custom ``__setattr__`` and speedup even more.