--- a/tests/test_functional.py
+++ b/tests/test_functional.py
@@ -747,6 +747,7 @@ class TestFunctional:
 
         src = inspect.getsource(D.__init__)
 
-        assert "_setattr(self, 'x', x)" in src
-        assert "_setattr(self, 'y', y)" in src
+        assert "_setattr = _cached_setattr_get(self)" in src
+        assert "_setattr('x', x)" in src
+        assert "_setattr('y', y)" in src
         assert object.__setattr__ != D.__setattr__