@davfsa 2022-08-02
Also worth noting that frozen classes don't scale well when it comes to attribute setting. When comparing setting 3 attributes in frozen and unfrozen slotted classes, you get a difference of about 200 nanoseconds. When you expand the test to 10 attributes, you get a difference of almost 500! (943 vs 445). 

Looks like the performance of making so many calls is not ideal, but I can't think of a better way of doing this without really stretching the limits of python.